@@ -32,7 +32,8 @@
     "plotly.js-dist-min": "^2.5.1",
     "rql": "github:epigeny/rql#1.1.0",
     "simplemde": "^1.11.2",
-    "vue": "^2.6.11"
+    "vue": "^3.4.0",
+    "@vue/compat": "^3.4.0"
   },
   "scripts": {
     "build:css": "sass --quiet-deps --style=compressed --source-map --load-path=node_modules src/main/resources/scss/mica.scss src/main/webapp/assets/css/mica.min.css",
